Fix: orderDto 변수 타입으로 인한 버그 수정

+++ src/main/java/com/myShop/order/OrderController.java
@@ -14,8 +14,8 @@ public class OrderController {
     private final OrderService service;
 
     @GetMapping("/get-order-id")
-    public ResponseEntity<Integer> getOrderId() {
-        int result = -1;
+    public ResponseEntity<Long> getOrderId() {
+        long result = -1;
         try {
             result = service.getOrderId();
             return new ResponseEntity<>(result, HttpStatus.OK);
