+++ src/main/java/com/myShop/order/OrderService.java
@@ -1,6 +1,6 @@
 package com.myShop.order;
 
 public interface OrderService {
-    public int getOrderId();
+    public long getOrderId();
     public boolean insert(OrderDto dto);
 }
